+++ src/main/java/roomescape/presentation/auth/dto/MemberRequestDto.java
@@ -1,4 +1,4 @@
-package roomescape.presentation.member.dto;
+package roomescape.presentation.auth.dto;
 
 import jakarta.validation.constraints.NotBlank;
