+++ src/test/java/roomescape/presentation/admin/api/AdminReservationControllerTest.java
@@ -26,7 +26,7 @@ import roomescape.persistence.fakerepository.FakeReservationThemeRepository;
 import roomescape.persistence.fakerepository.FakeReservationTimeRepository;
 import roomescape.presentation.AbstractControllerTest;
 import roomescape.presentation.admin.dto.AdminReservationRequestDto;
-import roomescape.presentation.member.dto.LoginRequestDto;
+import roomescape.presentation.auth.dto.LoginRequestDto;
 import roomescape.presentation.member.dto.ReservationResponseDto;
 
 class AdminReservationControllerTest extends AbstractControllerTest {
