+++ src/test/java/roomescape/presentation/admin/api/AdminReservationTimeControllerTest.java
@@ -12,8 +12,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.web.server.LocalServerPort;
-import roomescape.business.domain.member.Member;
-import roomescape.business.domain.member.MemberRole;
 import roomescape.business.domain.member.SignUpMember;
 import roomescape.business.domain.reservation.ReservationTime;
 import roomescape.exception.ErrorResponseDto;
@@ -22,7 +20,7 @@ import roomescape.persistence.fakerepository.FakeReservationTimeRepository;
 import roomescape.presentation.AbstractControllerTest;
 import roomescape.presentation.admin.dto.ReservationTimeRequestDto;
 import roomescape.presentation.admin.dto.ReservationTimeResponseDto;
-import roomescape.presentation.member.dto.LoginRequestDto;
+import roomescape.presentation.auth.dto.LoginRequestDto;
 
 class AdminReservationTimeControllerTest extends AbstractControllerTest {
